+++ app/src/main/java/me/capcom/smsgateway/modules/receiver/StateStorage.kt
@@ -14,10 +14,15 @@ class StateStorage(
         get() = storage.get<Long>(SMS_LAST_PROCESSED_ID) ?: 0
         set(value) = storage.set(SMS_LAST_PROCESSED_ID, value)
 
+    var sentSmsLastProcessedID: Long
+        get() = storage.get<Long>(SENT_SMS_LAST_PROCESSED_ID) ?: 0
+        set(value) = storage.set(SENT_SMS_LAST_PROCESSED_ID, value)
+
     companion object {
         private val PREFIX = "state."
 
         private val MMS_LAST_PROCESSED_ID = PREFIX + "last_processed_id"
         private val SMS_LAST_PROCESSED_ID = PREFIX + "sms_last_processed_id"
+        private val SENT_SMS_LAST_PROCESSED_ID = PREFIX + "sent_sms_last_processed_id"
     }
 }